+++ geopandas/geodataframe.py
@@ -87,9 +87,13 @@ class GeoDataFrame(GeoPandasBase, DataFrame):
         Coordinate Reference System of the geometry objects. Can be anything accepted by
         :meth:`pyproj.CRS.from_user_input() <pyproj.crs.CRS.from_user_input>`,
         such as an authority string (eg "EPSG:4326") or a WKT string.
-    geometry : str or array (optional)
-        If str, column to use as geometry. If array, will be set as 'geometry'
-        column on GeoDataFrame.
+    geometry : str or array-like (optional)
+        If str, column to use as active geometry column. If array-like, it will be
+        added as new column named 'geometry' column on GeoDataFrame and set as the
+        active geometry column. Note that if ``geometry`` is a (Geo)Series with a
+        name, the name will not be used, a column named "geometry" will still be
+        added. To preserve the name, you can use :meth:`~GeoDataFrame.rename_geometry`
+        to update the geometry column name.
 
     Examples
     --------
@@ -196,19 +200,7 @@ class GeoDataFrame(GeoPandasBase, DataFrame):
 
             if hasattr(geometry, "name") and geometry.name not in ("geometry", None):
                 # __init__ always creates geometry col named "geometry"
-                # but `set_geometry` respects the given series name
-                msg = (
-                    f"You have passed a Series named {geometry.name!r} as the active "
-                    f"geometry column to use, but the name {geometry.name!r} will be "
-                    "ignored as the GeoDataFrame constructor "
-                    'always creates a geometry column named "geometry". To silence '
-                    "this warning you can either call "
-                    'GeoDataFrame(..., geometry=ser.rename("geometry")) to keep the '
-                    'geometry column named "geometry", or you can call '
-                    "df.set_geometry(ser) to construct a GeoDataFrame with an active "
-                    f"geometry column called {geometry.name!r}."
-                )
-                warnings.warn(msg, category=UserWarning, stacklevel=2)
+                # rename as `set_geometry` respects the given series name
                 geometry = geometry.rename("geometry")
 
             self.set_geometry(geometry, inplace=True, crs=crs)